Streamlit/requirements.txt
numpy==1.26.2
seaborn==0.13.0
matplotlib==3.8.2
tensorflow==2.15.0
tensorflow-hub==0.15.0
scikit-learn==1.3.2
streamlit==1.24.0
pillow